[RESTEASY-3081] Move the CookieUtil to the test environment only.
https://issues.redhat.com/browse/RESTEASY-3081
