+++ resteasy-client/src/main/java/org/jboss/resteasy/client/jaxrs/internal/ClientRequestHeaders.java
@@ -1,7 +1,7 @@
 package org.jboss.resteasy.client.jaxrs.internal;
 
+import org.jboss.resteasy.spi.ResteasyProviderFactory;
 import org.jboss.resteasy.util.CaseInsensitiveMap;
-import org.jboss.resteasy.util.CookieUtil;
 import org.jboss.resteasy.util.DateUtil;
 import org.jboss.resteasy.util.HeaderHelper;
 import org.jboss.resteasy.util.MediaTypeHelper;
@@ -309,7 +309,9 @@ public class ClientRequestHeaders
          else
          {
             String str = configuration.toHeaderString(obj);
-            Cookie cookie = CookieUtil.valueOf(Cookie.class, str);
+            Cookie cookie = ResteasyProviderFactory.getInstance()
+                    .createHeaderDelegate(Cookie.class)
+                    .fromString(str);
             cookies.put(cookie.getName(), cookie);
          }
       }
