+++ testsuite/integration-tests/src/test/java/org/jboss/resteasy/test/resource/param/CookieInjectionTest.java
@@ -7,6 +7,7 @@ import org.jboss.resteasy.client.jaxrs.ProxyBuilder;
 import org.jboss.resteasy.client.jaxrs.ResteasyClientBuilder;
 import org.jboss.resteasy.spi.HttpResponseCodes;
 import org.jboss.resteasy.test.resource.param.resource.CookieInjectionResource;
+import org.jboss.resteasy.utils.CookieUtil;
 import org.jboss.resteasy.utils.PortProviderUtil;
 import org.jboss.resteasy.utils.TestUtil;
 import org.jboss.shrinkwrap.api.Archive;
@@ -59,7 +60,8 @@ public class CookieInjectionTest {
 
    @Deployment
    public static Archive<?> deploy() throws Exception {
-      WebArchive war = TestUtil.prepareArchive(CookieInjectionTest.class.getSimpleName());
+      WebArchive war = TestUtil.prepareArchive(CookieInjectionTest.class.getSimpleName())
+              .addClass(CookieUtil.class);
       return TestUtil.finishContainerPrepare(war, null, CookieInjectionResource.class);
    }
